+++ crewai_tools/tools/__init__.py
@@ -1,4 +1,4 @@
-from .ai_minds_tool.ai_minds_tool import AIMindTool
+from .ai_mind_tool.ai_mind_tool import AIMindTool
 from .brave_search_tool.brave_search_tool import BraveSearchTool
 from .browserbase_load_tool.browserbase_load_tool import BrowserbaseLoadTool
 from .code_docs_search_tool.code_docs_search_tool import CodeDocsSearchTool
